Fixed addon behavior when choosing to save settings globally
The old addon behavior was to keep two sets of settings, one global and one per character, so when the option to save settings globally was toggled on or off, the rest of the settings also ended up reverting to the corresponding version. This could be confusing from the player's perspective, so now only what you see in the active settings menu is stored and other saved versions are never reloaded.