Added FancyActionBar+ as an optional dependency
This change assumes that if anyone wants to use both CustomAbilityIcons and FancyActionBar+, they're trying to do so because they want the icons from CustomAbilityIcons (the more specific addon), with the extra functionality of FancyActionBar+ (the one with the most features). Since both addons utilize the same events, the easiest way to do this is to ensure that FancyActionBar+ loads before CustomAbilityIcons, which can be done by setting it as an optional dependency.